Move tests for cpu variants (v1,v2) to own module
