+++ crates/integration_test/Cargo.toml
@@ -27,4 +27,5 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 nix = "0.23.0"
 libcontainer = { path = "../libcontainer" }
+libcgroups = { path = "../libcgroups" }
 num_cpus = "1.0"
