+++ crates/integration_test/src/main.rs
@@ -63,9 +63,11 @@ fn main() -> Result<()> {
     let pidfile = get_pidfile_test();
     let ns_itype = get_ns_itype_tests();
     let cgroup_v1_pids = cgroups::pids::get_test_group();
-    let cgroup_v1_cpus = cgroups::cpus::get_test_group();
+    let cgroup_v1_cpu = cgroups::cpu::v1::get_test_group();
+    let cgroup_v2_cpu = cgroups::cpu::v2::get_test_group();
     let cgroup_v1_memory = cgroups::memory::get_test_group();
     let seccomp_notify = get_seccomp_notify_test();
+
 
     tm.add_test_group(&cl);
     tm.add_test_group(&cc);
@@ -73,11 +75,12 @@ fn main() -> Result<()> {
     tm.add_test_group(&pidfile);
     tm.add_test_group(&ns_itype);
     tm.add_test_group(&cgroup_v1_pids);
-    tm.add_test_group(&cgroup_v1_cpus);
+    tm.add_test_group(&cgroup_v1_cpu);
+    tm.add_test_group(&cgroup_v2_cpu);
     tm.add_test_group(&cgroup_v1_memory);
+    tm.add_test_group(&seccomp_notify);
 
     tm.add_cleanup(Box::new(cgroups::cleanup));
-    tm.add_test_group(&seccomp_notify);
 
     if let Some(tests) = opts.tests {
         let tests_to_run = parse_tests(&tests);
