+++ crates/integration_test/src/tests/cgroups/cpu/v2.rs
@@ -1,0 +1,9 @@
+use test_framework::TestGroup;
+
+fn can_run() -> bool {
+    todo!();
+}
+
+pub fn get_test_group<'a>() -> TestGroup<'a> {
+    todo!()
+}
